+++ src/react-components/CHANGELOG.md
@@ -12,6 +12,8 @@ Currently, this repo is in Prerelease.  When it is released, this project will a
 - `Icon` component that uses the `design-system-icons` package.
 - `SectionTitle` and `Heading` for heading components
 - `IconLink-List`for the list of subject links 
+- `Button` and `TextField` for the searchbar 
+- `SearchBar` component
 
 ### Changed 
 - `Breadcrumb` now shows icon in mobile view