+++ src/react-components/CHANGELOG.md
@@ -5,6 +5,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 Currently, this repo is in Prerelease.  When it is released, this project will adhere to [Semantic Versioning](https://semver.org/spec/v2.0.0.html)
 
 ========
+## [0.0.5] - 2019-12-06
+### Changed
+- `Icon` import
+
 ## [0.0.4] - 2019-12-06
 ### Added
 - `Body text` component for plain-text elements